Fix command quoting issues
The command the container is supposed to run is being passed to
pipes.quote() once in get_docker_run_cmd() and once again in
run_docker_container().

If the string contains anything that pipes.quote thinks should be
quoted then we end up with some very weird results. For example
"echo 'Hello' && sleep 60" turns to "echo foo '"'"'&&'"'"' sleep 60".

The fix here is to parse the whole docker command and comment only the
bits that passed to the container.

The initial code also prepended the commands with 'sh -c' which could
cause issues with the wrong quoting in some cases, so I also changed
the logic to look for /bin/sh (or /bin/bash or /bin/zsh) and only quote
the actual command instead of the whole thing.